Bug fix - Added a small function you can use
Bug fix for line 59   -->   it was trying to get the name of the player by using the owner of a territory, but commanders can occur (although it is unlikely and shouldn't happen) on neutral territories. 
This function (line 122) will give you the number of commanders in an Armies object. If the output of this function is 0, there are no commanders in the Armies object. If it is bigger than 0 there is at least one commander.